Add shipping column and modify moneytary value
change moneytary value to decimal and modify place order to include shipping fee

+++ Domain/Domain/Entities/BoxOption.cs
@@ -12,9 +12,9 @@ namespace Domain.Domain.Entities
         [Required]
         public string BoxOptionName { get; set; }  // nvarchar(200)
 
-        public float OriginPrice { get; set; }  // float
+        public decimal OriginPrice { get; set; }  // float
 
-        public float DisplayPrice { get; set; }  // float
+        public decimal DisplayPrice { get; set; }  // float
 
         public int BoxOptionStock { get; set; }  // int
         public bool IsDeleted { get; set; } = false;
