+++ Domain/Domain/Model/BoxOptionDTOs/AddBoxOptionDTO.cs
@@ -4,8 +4,8 @@
     {
         public string BoxOptionName { get; set; }  // nvarchar(200)
         public int BoxOptionStock { get; set; }  // int
-        public float OriginPrice { get; set; }  // float
-        public float DisplayPrice { get; set; }  // float
+        public decimal OriginPrice { get; set; }  // float
+        public decimal DisplayPrice { get; set; }  // float
         public bool IsDeleted { get; set; } = false;
         public int BoxId { get; set; }  // Foreign Key to Box
