+++ Domain/Domain/Model/BoxOptionDTOs/GetAllBoxOptionDTO.cs
@@ -7,8 +7,8 @@ namespace Domain.Domain.Model.BoxOptionDTOs
 
         public int BoxOptionId { get; set; }  // Primary Key
         public string BoxOptionName { get; set; }  // nvarchar(200)
-        public float OriginPrice { get; set; }  // float
-        public float DisplayPrice { get; set; }  // float
+        public decimal OriginPrice { get; set; }  // float
+        public decimal DisplayPrice { get; set; }  // float
         public int BoxOptionStock { get; set; }  // int
         public bool IsDeleted { get; set; } = false;
         public BelongBoxResponseDTO BelongBox { get; set; }
