+++ Domain/Domain/Model/OrderDTOs/Request/CreateOrderDtoDetail.cs
@@ -3,7 +3,7 @@
     public class CreateOrderDtoDetail
     {
         public CreateOrderDTO createOrderDto { get; set; }
-        public float revenue { get; set; }
+        public decimal revenue { get; set; }
         public bool openRequest { get; set; }
         public int currentOrderStatusId { get; set; }
         public string paymentStatus { get; set; }
