+++ Domain/Domain/Model/OrderItem/Request/OrderItemRequestDto.cs
@@ -3,9 +3,9 @@
     public class OrderItemRequestDto
     {
         public int quantity { get; set; }
-        public float price { get; set; }
+        public decimal price { get; set; }
         public int boxOptionId { get; set; }
-        public int originPrice { get; set; }
+        public decimal originPrice { get; set; }
         public int orderItemOpenRequestNumber { get; set; }
     }
 }
